pyproject.toml
[tool.poetry]
name = "mp4-to-wav-audio"
version = "0.1.0"
description = "Extracts .wav from a .MP4"
authors = ["Ryan Davis <ryancraigdavis@github.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ffmpeg-python = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
